LayoutTests/platform/win/fast/css/content-visibility-zoom-expected.txt

PASS lastSkipped is false
FAIL lastSkipped should be true. Was false.
FAIL lastSkipped should be true. Was false.
PASS successfullyParsed is true
Some tests failed.

TEST COMPLETE
